Adding more Filters and data validation

+++ api/Controllers/StockController.cs
@@ -1,4 +1,5 @@
 using api.Dto.Stock;
+using api.Helpers;
 using api.Interface;
 using api.Mapper;
 using Microsoft.AspNetCore.Mvc;
@@ -20,18 +21,25 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAllStock()
+        public async Task<IActionResult> GetAllStock([FromQuery] QueryObject queryObject)
         {
-            var stocks = await _stockRepo.GetAllStockAsync();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stocks = await _stockRepo.GetAllStockAsync(queryObject);
 
             var stockdtos = stocks.Select(s => s.toStockDto());
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        //Route Contstraints
+        [HttpGet("{id:int}")]
         //Model Binding
         public async Task<IActionResult> GetSingleStock([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var singleStock = await _stockRepo.GetSingleStockAsync(id);
 
             if (singleStock == null)
@@ -45,15 +53,21 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateStock([FromBody] CreateStockRequestDto createStock)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = createStock.toCreateRequestDto();
             await _stockRepo.CreateTheStockAsync(stockModel);
 
             return CreatedAtAction(nameof(GetSingleStock), new { id = stockModel.StockID }, stockModel.toStockDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> UpdateStock([FromRoute] int id, [FromBody] UpdateStockDto updateStock)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockToUpdate = await _stockRepo.UpdateTheStockAsync(id, updateStock);
             if (stockToUpdate == null)
             {
@@ -64,16 +78,19 @@ namespace api.Controllers
 
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> RemoveStock([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockToRemove = await _stockRepo.DeleteStockAsync(id);
             if (stockToRemove == null)
             {
                 return NotFound();
             }
 
-            return await GetAllStock();
+            return await GetAllStock(null);
             //return NoContent();
             //return NoContent() does same thing returns 204
